+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Implemented enhancements:**
 
+- Ignore docker network interface when running provisioning scripts [\#72](https://github.com/ferrarimarco/home-lab/issues/72)
 - Move Gateway configuration in a separate script [\#68](https://github.com/ferrarimarco/home-lab/issues/68)
 - Upgrade to docker-home-lab-dnsmasq:1.0.0 [\#66](https://github.com/ferrarimarco/home-lab/issues/66)
 - Configure Windows network zones [\#64](https://github.com/ferrarimarco/home-lab/issues/64)